install package passport

+++ database/migrations/2020_04_10_175646_create_products_table.php
@@ -17,7 +17,7 @@ class CreateProductsTable extends Migration
             $table->Increments('id');
             $table->string('name');
             $table->decimal('price');
-            $table->decimal('stock');
+            $table->integer('stock');
             $table->decimal('discount');
             $table->text('details');
             $table->timestamps();
